refactor: move getIsInstallable() to separate file
